+++ Tests.FMSystems.WeatherForecast.Api/Controllers/CityControllerTests.cs
@@ -1,6 +1,4 @@
-﻿using FMSystems.WeatherForecast.Api.Controllers;
-using FMSystems.WeatherForecast.Api.Controllers.City;
-using FMSystems.WeatherForecast.Domain.Entity;
+﻿using FMSystems.WeatherForecast.Api.Controllers.Cities;
 using FMSystems.WeatherForecast.Domain.Repository;
 using Microsoft.Extensions.Logging;
 using Moq;
